hwpxlib_python/requirements.txt
JPype1>=1.4.1